Generalize Twitch channel connection code
Previously, the code assumed that the bot should join the production
instance's Twitch channel on startup. It also assumed that the valid
channel for the `!join` and `!part` commands would only be the 9kmmrbot
channel.

The code should be deployable to create testing/staging instances of the
bot, so the above assumptions no longer hold true.